Final changes, making yyleng a size_t type

+++ include/c_lexer.hpp
@@ -19,7 +19,7 @@ extern std::string *yylval;
 // flex function to run on input
 extern int yylex();
 
-extern int yyleng;
+extern size_t yyleng;
 
 extern int lineCount;
 extern int spaceCount;
